requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.13
langchain-huggingface>=0.0.3
huggingface-hub>=0.20.3
flask>=2.0.1
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.22.0
openpyxl>=3.1.2
tiktoken>=0.5.1
redis>=5.0.1
hnswlib>=0.7.0
pillow>=10.0.0
pydantic>=2.5.2
faiss-cpu>=1.7.4
flask-cors>=4.0.0
langchainhub>=0.1.13
sentence-transformers>=2.2.2
tqdm>=4.66.1 